// src/interfaces/applicationWide.slice.interface.ts
// [NOTE]: Not currently used, don't know if needed. @Samir decide if you want to use or not and remove if not needed.

export interface IInitalStateApplicationWide {
  message?: string
  secondaryMessage?: string
  bgColor?: string
  borderColor?: string
  textColor?: string
  show?: boolean
  fromTop?: string
  status?: string
}